# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OrganiserProfileInput {
  bio: string;
  instagram: string;
  twitter: string;
}

export type OrganiserProfile = OrganiserProfileInput;

export interface SignInInput {
  email: string;
  /**
   * @minLength 8
   * @maxLength 72
   */
  password: string;
}

export type SignUpInputRole =
  (typeof SignUpInputRole)[keyof typeof SignUpInputRole];

export const SignUpInputRole = {
  user: "user",
  organizer: "organizer",
} as const;

export interface SignUpInput {
  email: string;
  /**
   * @minLength 8
   * @maxLength 72
   */
  password: string;
  /**
   * @minLength 2
   * @maxLength 120
   */
  displayName: string;
  role: SignUpInputRole;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  user: "user",
  organizer: "organizer",
} as const;

export interface AuthUser {
  id: string;
  name: string;
  email: string;
  profilePicture: string;
  isOrganizer: boolean;
  isAdmin: boolean;
  role: AuthUserRole;
  isVerified: boolean;
}