# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from "./authUserRole";

export interface AuthUser {
  id: string;
  name: string;
  email: string;
  profilePicture: string;
  isOrganizer: boolean;
  isAdmin: boolean;
  role: AuthUserRole;
  isVerified: boolean;
}