# lib/api-zod/src/generated/types/organiserProfile.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrganiserProfileInput } from "./organiserProfileInput";

export type OrganiserProfile = OrganiserProfileInput;