# lib/api-zod/src/generated/types/signInInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignInInput {
  email: string;
  /**
   * @minLength 8
   * @maxLength 72
   */
  password: string;
}