# lib/api-zod/src/generated/types/signUpInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignUpInputRole } from "./signUpInputRole";

export interface SignUpInput {
  email: string;
  /**
   * @minLength 8
   * @maxLength 72
   */
  password: string;
  /**
   * @minLength 2
   * @maxLength 120
   */
  displayName: string;
  role: SignUpInputRole;
}